Solution for sidebar, and getting image in dish
Could not solve full-height sidebar wrapper. In lab4 we "cheated" by
loading all images beforhand and therefore create border correct. Now
that ain't a solution because we aren't loading anything at start.

The solution is to have it as it is.

This commit also includes getting an image in dish.html via routing.
Next step will be to build that layout (for dish.html).